sentiment_server/requirements.txt
transformers
datasets
torch
scikit-learn
openai-whisper
flask